Allow `next build -F next` to be run when offline
With `NEXT_TEST_PREFER_OFFLINE=1 pnpm test...` we can already run
internal Next.js tests when being offline.

This PR expands on that by ensuring that no network requests are
triggered when using `NEXT_TEST_PREFER_OFFLINE=1 next build -F next`.

Specifically, the download of the AMP validator (`validator_wasm.js`) is
skipped, if the environment variable is set. To make this work, we're
skipping the deletion of the `packages/next/dist` directory, which is
normally done at the start of the build process, and reuse the
previously built files, including the AMP validator.